+++ src/RunGmicProgressChangedEventArgs.cs
@@ -10,30 +10,22 @@
 ////////////////////////////////////////////////////////////////////////
 
 using System;
+using System.ComponentModel;
 
 namespace GmicSharp
 {
     /// <summary>
-    /// The <see cref="EventArgs"/> used by the <see cref="Gmic{TGmicBitmap}.GmicProgress"/> event.
+    /// The <see cref="EventArgs"/> used by the <see cref="Gmic{TGmicBitmap}.RunGmicProgressChanged"/> event.
     /// </summary>
-    /// <seealso cref="EventArgs" />
-    public sealed class GmicProgressEventArgs : EventArgs
+    /// <seealso cref="ProgressChangedEventArgs" />
+    public sealed class RunGmicProgressChangedEventArgs : ProgressChangedEventArgs
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="GmicProgressEventArgs"/> class.
+        /// Initializes a new instance of the <see cref="RunGmicProgressChangedEventArgs"/> class.
         /// </summary>
         /// <param name="progress">The progress.</param>
-        public GmicProgressEventArgs(int progress)
+        public RunGmicProgressChangedEventArgs(int progress) : base(progress, null)
         {
-            Progress = progress;
         }
-
-        /// <summary>
-        /// Gets the progress.
-        /// </summary>
-        /// <value>
-        /// The progress.
-        /// </value>
-        public int Progress { get; }
     }
 }