+++ CHANGELOG.md
@@ -15,10 +15,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  * The `CustomResourcePath` property now supports paths non-ASCII characters.
  * Throw an exception if `AddInputImage` is called while G'MIC is running.
+ * Updated Gmic<TGmicBitmap>` to conform to the [Event-based Asynchronous Pattern](https://docs.microsoft.com/en-us/dotnet/standard/asynchronous-programming-patterns/event-based-asynchronous-pattern-overview).
 
 ### Removed
 
  * The `CustomUserFilePath` property in  `Gmic<TGmicBitmap>`.
+ * The `RunGmic(string, CancellationToken)` method.
 
 ## v0.6.0